+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Sandbox Apps: Shows now the number of loaded resources, the resources size and (if available) the memory usage of the page
  * Mashroom Portal: The App Info shows now also the number of the loaded resources for an App and the decoded size of those resources
  * Core: Use [nx](https://nx.dev) for building in dev mode if available. This should lead to a much faster startup in dev mode,
    especially if the distributed cloud cache is used.